Moves Criterion to Filter/Criterion
