+++ Tests/Filter/Criterion/CriterionTest.php
@@ -3,7 +3,7 @@
 namespace Ucc\Tests\Filter;
 
 use \PHPUnit_Framework_TestCase as TestCase;
-use Ucc\Filter\Criterion;
+use Ucc\Filter\Criterion\Criterion;
 
 class CriterionTest extends TestCase
 {
@@ -16,10 +16,10 @@ class CriterionTest extends TestCase
     public function testSetLogic()
     {
         $criterion = new Criterion();
-        $this->assertInstanceOf('Ucc\Filter\Criterion', $criterion->setLogic(Criterion::CRITERION_LOGIC_INTERSCTION));
+        $this->assertInstanceOf('Ucc\Filter\Criterion\Criterion', $criterion->setLogic(Criterion::CRITERION_LOGIC_INTERSCTION));
         $this->assertSame(Criterion::CRITERION_LOGIC_INTERSCTION, $criterion->logic());
 
-        $this->assertInstanceOf('Ucc\Filter\Criterion', $criterion->setLogic('OR'));
+        $this->assertInstanceOf('Ucc\Filter\Criterion\Criterion', $criterion->setLogic('OR'));
         $this->assertSame(Criterion::CRITERION_LOGIC_UNION, $criterion->logic());
     }
 
@@ -29,14 +29,14 @@ class CriterionTest extends TestCase
     public function testSetLogicFail()
     {
         $criterion = new Criterion();
-        $this->assertInstanceOf('Ucc\Filter\Criterion', $criterion->setLogic('abc'));
+        $this->assertInstanceOf('Ucc\Filter\Criterion\Criterion', $criterion->setLogic('abc'));
     }
 
     public function testSetKey()
     {
         $criterion  = new Criterion();
         $expected   = 'email';
-        $this->assertInstanceOf('Ucc\Filter\Criterion', $criterion->setKey($expected));
+        $this->assertInstanceOf('Ucc\Filter\Criterion\Criterion', $criterion->setKey($expected));
         $this->assertSame($expected, $criterion->key());
     }
 
@@ -44,7 +44,7 @@ class CriterionTest extends TestCase
     {
         $criterion  = new Criterion();
         $expected   = 'eq';
-        $this->assertInstanceOf('Ucc\Filter\Criterion', $criterion->setOperand($expected));
+        $this->assertInstanceOf('Ucc\Filter\Criterion\Criterion', $criterion->setOperand($expected));
         $this->assertSame($expected, $criterion->op());
     }
 
@@ -55,14 +55,14 @@ class CriterionTest extends TestCase
     {
         $criterion  = new Criterion();
         $expected   = 'foo';
-        $this->assertInstanceOf('Ucc\Filter\Criterion', $criterion->setOperand($expected));
+        $this->assertInstanceOf('Ucc\Filter\Criterion\Criterion', $criterion->setOperand($expected));
     }
 
     public function testSetType()
     {
         $criterion  = new Criterion();
         $expected   = 'value';
-        $this->assertInstanceOf('Ucc\Filter\Criterion', $criterion->setType($expected));
+        $this->assertInstanceOf('Ucc\Filter\Criterion\Criterion', $criterion->setType($expected));
         $this->assertSame($expected, $criterion->type());
     }
 
@@ -73,14 +73,14 @@ class CriterionTest extends TestCase
     {
         $criterion  = new Criterion();
         $expected   = 'foo';
-        $this->assertInstanceOf('Ucc\Filter\Criterion', $criterion->setType($expected));
+        $this->assertInstanceOf('Ucc\Filter\Criterion\Criterion', $criterion->setType($expected));
     }
 
     public function testSetValue()
     {
         $criterion  = new Criterion();
         $expected   = 123;
-        $this->assertInstanceOf('Ucc\Filter\Criterion', $criterion->setvalue($expected));
+        $this->assertInstanceOf('Ucc\Filter\Criterion\Criterion', $criterion->setvalue($expected));
         $this->assertSame($expected, $criterion->value());
     }
 }
